requirements.txt
Click==7.0
Flask==1.1.1
Flask-Session==0.3.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
pkg-resources==0.0.0
Werkzeug==0.16.0
WTForms==2.2.1
